backend/requirements.txt
pip==25.0.1
fastapi
uvicorn
httpx
pydantic
python-dotenv
vital
pytest
